cast store argument to proper uint

+++ unittest/data/codegen-shader/compute_buffer_write.sabre.out.hlsl
@@ -10,11 +10,11 @@ void main_main() {
 	int _tmp_1 = int(main_numbers.Load(32));
 	int _tmp_2 = int(main_numbers.Load(36));
 	int _tmp_3 = int(main_numbers.Load(40 + ((_tmp_1 + _tmp_2) * 4)));
-	main_numbers.Store(40 + ((9) * 4), _tmp_3);
+	main_numbers.Store(40 + ((9) * 4), uint(_tmp_3));
 	int _tmp_4 = int(main_numbers.Load(40 + ((9) * 4)));
-	main_numbers.Store(40 + ((9) * 4), _tmp_4 + 42);
+	main_numbers.Store(40 + ((9) * 4), uint(_tmp_4 + 42));
 	int _tmp_5 = int(main_numbers.Load(40 + ((9) * 4)));
-	main_numbers.Store(40 + ((9) * 4), _tmp_5 - 1);
+	main_numbers.Store(40 + ((9) * 4), uint(_tmp_5 - 1));
 }
 
 [numthreads(1, 1, 1)]
